refactor(execution-engine): extract task-runner-prepare from task-runner
